♻️ JE-103-refactor: date cell 컴포넌트화
